# gh-104635: Expand optimization for dead store elimination