<!--
Thanks for your contribution!
Please read this comment in its entirety. It's quite important.

# Pull Request title

It should be in the following format:

```
gh-NNNNN: Summary of the changes made
```

Where: gh-NNNNN refers to the GitHub issue number.

Most PRs will require an issue number. Trivial changes, like fixing a typo, do not need an issue.

# Backport Pull Request title

If this is a backport PR (PR made against branches other than `main`),
please ensure that the PR title is in the following format:

```
[X.Y] <title from the original PR> (GH-NNNN)
```

Where: [X.Y] is the branch name, e.g. [3.6].

GH-NNNN refers to the PR number from `main`.

-->

## Explain

> We might also want to expand this so it can look more than one instruction ahead for another write to the same location, skipping over an allowlist of instructions that we know don't read that same local, don't execute arbitrary code, and can't raise an exception. E.g. LOAD_FAST of a different local, POP_TOP, NOP, probably others.

This PR is motivated by @carljm 's comment from https://github.com/python/cpython/pull/105040#issuecomment-1570770763
The limitation of https://github.com/python/cpython/pull/105320 is that the optimization only covers the next redundant opcode, but with this patch, the optimization can be applied to more than one instruction ahead for another write to the same location.

## Benchmark

````python

import pyperf

runner = pyperf.Runner()
runner.timeit(name="bench dead_store",
              stmt="""
_, a, b, c, _ = 1, 2, 3, 4, 5
""")

````
````
Mean +- std dev: [base] 13.1 ns +- 0.2 ns -> [opt] 12.6 ns +- 0.2 ns: 1.04x faster
````

## TODO (Done)
- [x] Fix test_dis.py
- [x] Add testcase for this optimization.


<!-- gh-issue-number: gh-104635 -->
* Issue: gh-104635
<!-- /gh-issue-number -->