@corona10 2023-07-09
````

./_bootstrap_python ./Programs/_freeze_module.py runpy ./Lib/runpy.py Python/frozen_modules/runpy.h
Assertion failed: (OPCODE_HAS_ARG(opcode) || HAS_TARGET(opcode) || oparg == 0), function instr_sequence_addop, file compile.c, line 243.
make: *** [Python/frozen_modules/runpy.h] Abort trap: 6
make: *** Waiting for unfinished jobs....
````

Interesting crash; this crash was not triggered at my local machine. I will take a look.