@corona10 2023-07-10
Umm, this patch is missing the following case. (same line, multi variable multi target case)
I will improve for such case, do not merge yet :)

````

a, a, b, b = 1, 2, 3, 4
````