@corona10 2023-07-10
```
>>> def f():
...     a = 42; b = a + 54; a = 54
...     return a, b
...
>>> f()
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "<stdin>", line 2, in f
UnboundLocalError: cannot access local variable 'a' where it is not associated with a value
```

Found more edge case, I will close the PR. We may need some kinds of use-def analysis