@corona10 2023-07-10
> If you are only looking at linear sequences of instructions that are on a single line then, almost by definition, a superoptimizer will do at least as well.

Yeah, I agree if the super-optimizer is added, I expect that most of the cases will be covered, but I just tried it because the super-optimizer infra does not exist at this moment.

More digging will be more expensive than I expect,  so I am not going to cover those cases at this moment.

(At least, we may need to add use-def analysis, but it should be compared to adding a super-optimizer process)